+++ clx_task_management/models/sale_order.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo, CLx Media
 # See LICENSE file for full copyright & licensing details.
-from odoo import fields, models, api
+from odoo import fields, models, api, _
 from odoo.osv import expression
+from odoo.exceptions import UserError
 
 
 class SaleOrder(models.Model):
@@ -21,10 +22,14 @@ class SaleOrder(models.Model):
             if subscriptions:
                 active_subscription_lines = subscriptions.recurring_invoice_line_ids.filtered(
                     lambda x: (x.start_date and x.end_date and x.start_date <= today <= x.end_date)
-                    or (x.start_date and not x.end_date and x.start_date <= today)
-                    )
-            if active_subscription_lines:
-                sale_orders = active_subscription_lines.mapped('so_line_id').mapped('order_id')
+                              or (x.start_date and not x.end_date and x.start_date <= today)
+                )
+                future_subscription_lines = subscriptions.recurring_invoice_line_ids.filtered(
+                    lambda x: x.start_date and x.start_date >= today)
+                if active_subscription_lines:
+                    sale_orders = active_subscription_lines.mapped('so_line_id').mapped('order_id')
+                if future_subscription_lines:
+                    sale_orders += future_subscription_lines.mapped('so_line_id').mapped('order_id')
             if sale_orders:
                 domain = expression.AND(
                     [args or [], [('id', 'in', sale_orders.ids)]])
@@ -33,6 +38,10 @@ class SaleOrder(models.Model):
 
     def _action_confirm(self):
         result = super(SaleOrder, self)._action_confirm()
+        won_stage_id = self.env.ref('crm.stage_lead4')
+        if won_stage_id and self.opportunity_id and self.opportunity_id.stage_id.id != won_stage_id.id:
+            raise UserError(_("You will need WON stage of opportunity {} to Confirm the Sale order.").format(self.opportunity_id.name))
+
         for line in self.order_line:
             if line.task_id and not line.task_id.clx_sale_order_id and not line.task_id.clx_sale_order_line_id:
                 line.task_id.write(